remove loader2 and style restart button
